+++ src/main/java/com/room414/racingbets/dal/abstraction/dao/RaceDao.java
@@ -1,6 +1,5 @@
 package com.room414.racingbets.dal.abstraction.dao;
 
-import com.room414.racingbets.dal.abstraction.exception.DalException;
 import com.room414.racingbets.dal.domain.entities.Race;
 import com.room414.racingbets.dal.domain.enums.RaceStatus;
 
@@ -18,34 +17,34 @@ public interface RaceDao extends CrudDao<Long, Race> {
     /**
      * @return List of Races that scheduled on RacecourseDto with id == racecourse param
      */
-    List<Race> findByRacecourseId(RaceStatus status, long racecourse, long offset, long limit);
+    List<Race> findByRacecourseId(RaceStatus status, long racecourse, int offset, int limit);
 
     /**
      * @return Count of Races that scheduled on RacecourseDto with id == racecourse param
      */
-    long findByRacecourseIdCount(RaceStatus status, long racecourse);
+    int findByRacecourseIdCount(RaceStatus status, long racecourse);
 
 
     /**
      * @return List of Races that scheduled on RacecourseDto with name starts with racecourse param
      */
-    List<Race> findByRacecourse(RaceStatus status, String racecourse, long offset, long limit);
+    List<Race> findByRacecourse(RaceStatus status, String racecourse, int offset, int limit);
 
     /**
      * @return Count of Races that scheduled on RacecourseDto with name starts with racecourse param
      */
-    long findByRacecourseCount(RaceStatus status, String racecourse);
+    int findByRacecourseCount(RaceStatus status, String racecourse);
 
 
     /**
      * @return List of Races that scheduled on timestamp between begin and end param
      */
-    List<Race> findInTimestampDiapason(RaceStatus status, Timestamp begin, Timestamp end, long offset, long limit);
+    List<Race> findInTimestampDiapason(RaceStatus status, Timestamp begin, Timestamp end, int offset, int limit);
 
     /**
      * @return Count of Races that scheduled on timestamp between begin and end param
      */
-    long findInTimestampDiapasonCount(RaceStatus status, Timestamp begin, Timestamp end);
+    int findInTimestampDiapasonCount(RaceStatus status, Timestamp begin, Timestamp end);
 
 
     /**
@@ -53,14 +52,14 @@ public interface RaceDao extends CrudDao<Long, Race> {
      *      on RacecourseDto with id == racecourse param
      */
     List<Race> findInTimestampDiapasonOnRacecourse(
-            RaceStatus status, long racecourse, Timestamp begin, Timestamp end, long offset, long limit
+            RaceStatus status, long racecourse, Timestamp begin, Timestamp end, int offset, int limit
     );
 
     /**
      * @return Count of Races that scheduled on timestamp between begin and end param
      *      on RacecourseDto id == racecourse param
      */
-    long findInTimestampDiapasonOnRacecourseCount(
+    int findInTimestampDiapasonOnRacecourseCount(
             RaceStatus status, long racecourse, Timestamp begin, Timestamp end
     );
 
@@ -68,12 +67,12 @@ public interface RaceDao extends CrudDao<Long, Race> {
     /**
      * @return List of Races which names starts with namePart or empty list if no found.
      */
-    List<Race> findByNamePart(RaceStatus raceStatus, String namePart, long offset, long limit);
+    List<Race> findByNamePart(RaceStatus raceStatus, String namePart, int offset, int limit);
 
     /**
      * @return count of Races which names starts with namePart
      */
-    long findByNamePartCount(RaceStatus raceStatus, String namePart);
+    int findByNamePartCount(RaceStatus raceStatus, String namePart);
 
     boolean updateStatus(long id, RaceStatus status);
 }
