+++ src/main/java/com/room414/racingbets/dal/concrete/caching/dao/CachedParticipantDao.java
@@ -4,7 +4,6 @@ import com.room414.racingbets.dal.abstraction.cache.ParticipantCache;
 import com.room414.racingbets.dal.abstraction.dao.ParticipantDao;
 import com.room414.racingbets.dal.abstraction.infrastructure.Pair;
 import com.room414.racingbets.dal.concrete.caching.caffeine.base.CacheCrudDao;
-import com.room414.racingbets.dal.concrete.caching.caffeine.caches.CaffeineParticipantCache;
 import com.room414.racingbets.dal.domain.entities.Participant;
 
 import java.sql.Timestamp;
@@ -25,56 +24,56 @@ public class CachedParticipantDao extends CacheCrudDao<Participant> implements P
     }
 
     @Override
-    public List<Pair<Participant, Timestamp>> findByHorseId(long id, long offset, long limit) {
+    public List<Pair<Participant, Timestamp>> findByHorseId(long id, int offset, int limit) {
         final String key = String.format("find:horse:%d:%d:%d", id, limit, offset);
 
         return cache.getWhoAndWhenCached(key, () -> dao.findByHorseId(id, offset, limit));
     }
 
     @Override
-    public long findByHorseIdCount(long id) {
+    public int findByHorseIdCount(long id) {
         final String key = String.format("find:horse:count:%d", id);
 
         return cache.getCachedCount(key, () -> dao.findByHorseIdCount(id));
     }
 
     @Override
-    public List<Pair<Participant, Timestamp>> findByOwnerId(long id, long offset, long limit) {
+    public List<Pair<Participant, Timestamp>> findByOwnerId(long id, int offset, int limit) {
         final String key = String.format("find:owner:%d:%d:%d", id, limit, offset);
 
         return cache.getWhoAndWhenCached(key, () -> dao.findByOwnerId(id, offset, limit));
     }
 
     @Override
-    public long findByOwnerIdCount(long id) {
+    public int findByOwnerIdCount(long id) {
         final String key = String.format("find:owner:count:%d", id);
 
         return cache.getCachedCount(key, () -> dao.findByOwnerIdCount(id));
     }
 
     @Override
-    public List<Pair<Participant, Timestamp>> findByJockeyId(long id, long offset, long limit) {
+    public List<Pair<Participant, Timestamp>> findByJockeyId(long id, int offset, int limit) {
         final String key = String.format("find:jockey:%d:%d:%d", id, limit, offset);
 
         return cache.getWhoAndWhenCached(key, () -> dao.findByJockeyId(id, offset, limit));
     }
 
     @Override
-    public long findByJockeyIdCount(long id) {
+    public int findByJockeyIdCount(long id) {
         final String key = String.format("find:jockey:count:%d", id);
 
         return cache.getCachedCount(key, () -> dao.findByJockeyIdCount(id));
     }
 
     @Override
-    public List<Pair<Participant, Timestamp>> findByTrainerId(long id, long offset, long limit) {
+    public List<Pair<Participant, Timestamp>> findByTrainerId(long id, int offset, int limit) {
         final String key = String.format("find:trainer:%d:%d:%d", id, limit, offset);
 
         return cache.getWhoAndWhenCached(key, () -> dao.findByTrainerId(id, offset, limit));
     }
 
     @Override
-    public long findByTrainerIdCount(long id) {
+    public int findByTrainerIdCount(long id) {
         final String key = String.format("find:trainer:count:%d", id);
 
         return cache.getCachedCount(key, () -> dao.findByTrainerIdCount(id));
