+++ src/main/java/com/room414/racingbets/dal/concrete/caching/dao/CachedRaceDao.java
@@ -4,7 +4,6 @@ import com.room414.racingbets.dal.abstraction.cache.RaceCache;
 import com.room414.racingbets.dal.abstraction.dao.RaceDao;
 import com.room414.racingbets.dal.abstraction.exception.DalException;
 import com.room414.racingbets.dal.concrete.caching.caffeine.base.CacheCrudDao;
-import com.room414.racingbets.dal.concrete.caching.caffeine.caches.CaffeineRaceCache;
 import com.room414.racingbets.dal.domain.entities.Race;
 import com.room414.racingbets.dal.domain.enums.RaceStatus;
 
@@ -26,35 +25,35 @@ public class CachedRaceDao extends CacheCrudDao<Race> implements RaceDao {
     }
 
     @Override
-    public List<Race> findByRacecourseId(RaceStatus status, long racecourse, long offset, long limit) {
+    public List<Race> findByRacecourseId(RaceStatus status, long racecourse, int offset, int limit) {
         final String key = String.format("find:racecourse:id:%s:%d:%d:%d", status, racecourse, offset, limit);
 
         return cache.getManyCached(key, () -> dao.findByRacecourseId(status, racecourse, offset, limit));
     }
 
     @Override
-    public long findByRacecourseIdCount(RaceStatus status, long racecourse) throws DalException {
+    public int findByRacecourseIdCount(RaceStatus status, long racecourse) throws DalException {
         final String key = String.format("find:racecourse:id:count:%s:%d:", status, racecourse);
 
         return cache.getCachedCount(key, () -> dao.findByRacecourseIdCount(status, racecourse));
     }
 
     @Override
-    public List<Race> findByRacecourse(RaceStatus status, String racecourse, long offset, long limit) {
+    public List<Race> findByRacecourse(RaceStatus status, String racecourse, int offset, int limit) {
         final String key = String.format("find:racecourse:name:%s:%s:%d:%d", status, racecourse, offset, limit);
 
         return cache.getManyCached(key, () -> dao.findByRacecourse(status, racecourse, offset, limit));
     }
 
     @Override
-    public long findByRacecourseCount(RaceStatus status, String racecourse) {
+    public int findByRacecourseCount(RaceStatus status, String racecourse) {
         final String key = String.format("find:racecourse:name:count:%s:%s:", status, racecourse);
 
         return cache.getCachedCount(key, () -> dao.findByRacecourseCount(status, racecourse));
     }
 
     @Override
-    public List<Race> findInTimestampDiapason(RaceStatus status, Timestamp begin, Timestamp end, long offset, long limit) {
+    public List<Race> findInTimestampDiapason(RaceStatus status, Timestamp begin, Timestamp end, int offset, int limit) {
         final String key = String.format(
                 "find:timestamp:%s:%s:%s:%d:%d", status, begin, end, offset, limit
         );
@@ -63,14 +62,14 @@ public class CachedRaceDao extends CacheCrudDao<Race> implements RaceDao {
     }
 
     @Override
-    public long findInTimestampDiapasonCount(RaceStatus status, Timestamp begin, Timestamp end) {
+    public int findInTimestampDiapasonCount(RaceStatus status, Timestamp begin, Timestamp end) {
         final String key = String.format("find:timestamp:count:%s:%s:%s", status, begin, end);
 
         return cache.getCachedCount(key, () -> dao.findInTimestampDiapasonCount(status, begin, end));
     }
 
     @Override
-    public List<Race> findInTimestampDiapasonOnRacecourse(RaceStatus status, long racecourse, Timestamp begin, Timestamp end, long offset, long limit) {
+    public List<Race> findInTimestampDiapasonOnRacecourse(RaceStatus status, long racecourse, Timestamp begin, Timestamp end, int offset, int limit) {
         final String key = String.format(
                 "find:timestamp:%s:%d:%s:%s:%d:%d", status, racecourse, begin, end, offset, limit
         );
@@ -81,7 +80,7 @@ public class CachedRaceDao extends CacheCrudDao<Race> implements RaceDao {
     }
 
     @Override
-    public long findInTimestampDiapasonOnRacecourseCount(RaceStatus status, long racecourse, Timestamp begin, Timestamp end) {
+    public int findInTimestampDiapasonOnRacecourseCount(RaceStatus status, long racecourse, Timestamp begin, Timestamp end) {
         final String key = String.format("find:timestamp:count:%s:%d:%s:%s", status, racecourse, begin, end);
 
         return cache.getCachedCount(key, () -> dao.findInTimestampDiapasonOnRacecourseCount(
@@ -90,14 +89,14 @@ public class CachedRaceDao extends CacheCrudDao<Race> implements RaceDao {
     }
 
     @Override
-    public List<Race> findByNamePart(RaceStatus raceStatus, String namePart, long offset, long limit) {
+    public List<Race> findByNamePart(RaceStatus raceStatus, String namePart, int offset, int limit) {
         String key = String.format("find:name:%s:%s:%d:%d", raceStatus, namePart, limit, offset);
 
         return cache.getManyCached(key, () -> dao.findByNamePart(raceStatus, namePart, offset, limit));
     }
 
     @Override
-    public long findByNamePartCount(RaceStatus raceStatus, String namePart) {
+    public int findByNamePartCount(RaceStatus raceStatus, String namePart) {
         String key = String.format("find:name:count:%s:%s", raceStatus, namePart);
 
         return cache.getCachedCount(key, () -> dao.findByNamePartCount(raceStatus, namePart));
