+++ src/main/java/com/room414/racingbets/dal/concrete/caching/infrastructure/lazyload/LazyLoadRaceDao.java
@@ -2,7 +2,6 @@ package com.room414.racingbets.dal.concrete.caching.infrastructure.lazyload;
 
 import com.room414.racingbets.dal.abstraction.dao.RaceDao;
 import com.room414.racingbets.dal.abstraction.dao.UnitOfWork;
-import com.room414.racingbets.dal.abstraction.exception.DalException;
 import com.room414.racingbets.dal.domain.entities.Race;
 import com.room414.racingbets.dal.domain.enums.RaceStatus;
 
@@ -30,7 +29,7 @@ public class LazyLoadRaceDao implements RaceDao {
     }
 
     @Override
-    public List<Race> findByRacecourseId(RaceStatus status, long racecourse, long offset, long limit) {
+    public List<Race> findByRacecourseId(RaceStatus status, long racecourse, int offset, int limit) {
         return getRaceDao().findByRacecourseId(status, racecourse, offset, limit);
     }
 
@@ -45,27 +44,27 @@ public class LazyLoadRaceDao implements RaceDao {
     }
 
     @Override
-    public long findByRacecourseIdCount(RaceStatus status, long racecourse) {
+    public int findByRacecourseIdCount(RaceStatus status, long racecourse) {
         return getRaceDao().findByRacecourseIdCount(status, racecourse);
     }
 
     @Override
-    public List<Race> findByRacecourse(RaceStatus status, String racecourse, long offset, long limit) {
+    public List<Race> findByRacecourse(RaceStatus status, String racecourse, int offset, int limit) {
         return getRaceDao().findByRacecourse(status, racecourse, offset, limit);
     }
 
     @Override
-    public List<Race> findAll(long offset, long limit) {
+    public List<Race> findAll(int offset, int limit) {
         return getRaceDao().findAll(offset, limit);
     }
 
     @Override
-    public long count() {
+    public int count() {
         return getRaceDao().count();
     }
 
     @Override
-    public long findByRacecourseCount(RaceStatus status, String racecourse) {
+    public int findByRacecourseCount(RaceStatus status, String racecourse) {
         return getRaceDao().findByRacecourseCount(status, racecourse);
     }
 
@@ -75,7 +74,7 @@ public class LazyLoadRaceDao implements RaceDao {
     }
 
     @Override
-    public List<Race> findInTimestampDiapason(RaceStatus status, Timestamp begin, Timestamp end, long offset, long limit) {
+    public List<Race> findInTimestampDiapason(RaceStatus status, Timestamp begin, Timestamp end, int offset, int limit) {
         return getRaceDao().findInTimestampDiapason(status, begin, end, offset, limit);
     }
 
@@ -85,27 +84,27 @@ public class LazyLoadRaceDao implements RaceDao {
     }
 
     @Override
-    public long findInTimestampDiapasonCount(RaceStatus status, Timestamp begin, Timestamp end) {
+    public int findInTimestampDiapasonCount(RaceStatus status, Timestamp begin, Timestamp end) {
         return getRaceDao().findInTimestampDiapasonCount(status, begin, end);
     }
 
     @Override
-    public List<Race> findInTimestampDiapasonOnRacecourse(RaceStatus status, long racecourse, Timestamp begin, Timestamp end, long offset, long limit) {
+    public List<Race> findInTimestampDiapasonOnRacecourse(RaceStatus status, long racecourse, Timestamp begin, Timestamp end, int offset, int limit) {
         return getRaceDao().findInTimestampDiapasonOnRacecourse(status, racecourse, begin, end, offset, limit);
     }
 
     @Override
-    public long findInTimestampDiapasonOnRacecourseCount(RaceStatus status, long racecourse, Timestamp begin, Timestamp end) {
+    public int findInTimestampDiapasonOnRacecourseCount(RaceStatus status, long racecourse, Timestamp begin, Timestamp end) {
         return getRaceDao().findInTimestampDiapasonOnRacecourseCount(status, racecourse, begin, end);
     }
 
     @Override
-    public List<Race> findByNamePart(RaceStatus raceStatus, String namePart, long offset, long limit) {
+    public List<Race> findByNamePart(RaceStatus raceStatus, String namePart, int offset, int limit) {
         return getRaceDao().findByNamePart(raceStatus, namePart, offset, limit);
     }
 
     @Override
-    public long findByNamePartCount(RaceStatus raceStatus, String namePart) {
+    public int findByNamePartCount(RaceStatus raceStatus, String namePart) {
         return getRaceDao().findByNamePartCount(raceStatus, namePart);
     }
